+++ src/redirect.rs
@@ -17,7 +17,7 @@ pub fn redirect_legacy_illust(request: &Request) -> Response {
     Response::redirect_301(destination)
 }
 
-pub fn redirect_fanbox(client: &ureq::Agent, id: u32) -> Response {
+pub fn redirect_fanbox(client: &ureq::Agent, id: u64) -> Response {
     let url = format!("https://www.pixiv.net/fanbox/creator/{}", id);
     let response = client.get(&url).call().unwrap();
     let location = response.header("Location").unwrap().to_string();
